02-use-cases/asynchronous-shopping-assistant/requirements.txt
bedrock-agentcore-sdk
langgraph
langchain
langchain_aws
typing_extensions
langchain_community
nova_act
